Update to the TransitioningBackgrounds.plugin.js
Makes the plugin usable again fixes the obnoxious transparency issue of the plugin where you only can see the background behind the channels and not the actual chat. so now you can actually see your background now. also adds support for backgrounds on the new pages such as store, activity, library and make the guild bar transparent i personally like it transparent some might not Sue me this plugin seems to be abandoned since this problem has been a thing for over 4 to 5 months but heres the fix if people still want to use it.