Update DetailedServerTooltips.plugin.js
loading all members for servers with less than 500 members. this should be a good compromise between loading the correct member count (bc it seems like the module i changed to shows more users than there really are) and no useless ram usage

also not loading all users if the owner is missing anymore. only load the owner itself and save the tag to not have to load it again all the time